Initial API redesign
Initial implementation of the API redesign prompted by a suggestion by
@xiaq (https://github.com/zzamboni/elvish-completions/pull/6). Instead
of a single entry point comp:expand, which examines its data structure
to see what to do with it, the new API has multiple entry functions,
and each receive arguments according to their function. This makes the
overall structure of a complex completion definition cleaner and
easier.

In this first rewrite, only the outward-facing functions have been
updated, but internally the same old data structures are used. This is
meant to allow the existing completions to be rewritten, to get a feel
for and exercise the new API. A more thorough update will follow.